requirements_nl.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements_nl.in -o requirements_nl.txt
ago==0.0.95
    # via -r requirements_nl.in
awscrt==0.27.6
    # via botocore
blinker==1.9.0
    # via
    #   flask
    #   gds-metrics
    #   sentry-sdk
boto3==1.40.28
    # via notifications-utils
botocore==1.40.28
    # via
    #   boto3
    #   s3transfer
cachetools==6.2.0
    # via notifications-utils
certifi==2025.8.3
    # via
    #   requests
    #   sentry-sdk
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via flask
cryptography==44.0.3
    # via fido2
dnspython==2.8.0
    # via eventlet
docopt==0.6.2
    # via notifications-python-client
et-xmlfile==2.0.0
    # via openpyxl
eventlet==0.40.3
    # via gunicorn
fido2==1.1.3
    # via -r requirements_nl.in
flask==3.1.2
    # via
    #   flask-login
    #   flask-redis
    #   flask-wtf
    #   gds-metrics
    #   notifications-utils
    #   sentry-sdk
flask-login==0.6.3
    # via -r requirements_nl.in
flask-redis==0.4.0
    # via notifications-utils
flask-wtf==1.2.1
    # via -r requirements_nl.in
gds-metrics @ git+https://github.com/alphagov/gds_metrics_python.git@6f1840a57b6fb1ee40b7e84f2f18ec229de8aa72
    # via -r requirements_nl.in
govuk-bank-holidays==0.17
    # via notifications-utils
govuk-frontend-jinja @ git+https://github.com/Worth-NL/LandRegistry-frontend-jinja@0d6796c785db6db67ed3cf375df2dc3bc318c267
    # via -r requirements_nl.in
greenlet==3.2.4
    # via eventlet
gunicorn==23.0.0
    # via notifications-utils
humanize==4.4.0
    # via -r requirements_nl.in
idna==3.10
    # via requests
itsdangerous==2.2.0
    # via
    #   flask
    #   flask-wtf
    #   notifications-utils
jinja2==3.1.6
    # via
    #   flask
    #   govuk-frontend-jinja
    #   notifications-utils
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
lml==0.2.0
    # via
    #   pyexcel
    #   pyexcel-io
lxml==6.0.1
    # via
    #   pyexcel-ezodf
    #   pyexcel-ods3
markupsafe==3.0.2
    # via
    #   flask
    #   jinja2
    #   sentry-sdk
    #   werkzeug
    #   wtforms
mistune==0.8.4
    # via notifications-utils
notifications-python-client==10.0.0
    # via -r requirements_nl.in
notifications-utils @ git+https://github.com/Worth-NL/notifynl-utils.git@02eb2655d064aee09adc4ad3c68c663b1826ed75
    # via -r requirements_nl.in
openpyxl==3.1.5
    # via pyexcel-xlsx
ordered-set==4.1.0
    # via notifications-utils
packaging==25.0
    # via gunicorn
phonenumbers==9.0.13
    # via notifications-utils
pillow==10.3.0
    # via -r requirements_nl.in
prometheus-client==0.15.0
    # via
    #   -r requirements_nl.in
    #   gds-metrics
pycparser==2.23
    # via cffi
pyexcel==0.7.1
    # via -r requirements_nl.in
pyexcel-ezodf==0.3.4
    # via pyexcel-ods3
pyexcel-io==0.6.7
    # via
    #   -r requirements_nl.in
    #   pyexcel
    #   pyexcel-ods3
    #   pyexcel-xls
    #   pyexcel-xlsx
pyexcel-ods3==0.6.1
    # via -r requirements_nl.in
pyexcel-xls==0.7.0
    # via -r requirements_nl.in
pyexcel-xlsx==0.6.1
    # via -r requirements_nl.in
pyjwt==2.10.1
    # via notifications-python-client
pypdf==6.0.0
    # via notifications-utils
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   notifications-utils
python-json-logger==3.3.0
    # via notifications-utils
pytz==2025.2
    # via notifications-utils
pyyaml==6.0.2
    # via notifications-utils
redis==6.4.0
    # via flask-redis
requests==2.32.5
    # via
    #   govuk-bank-holidays
    #   notifications-python-client
    #   notifications-utils
s3transfer==0.14.0
    # via boto3
segno==1.6.6
    # via notifications-utils
sentry-sdk==1.45.1
    # via -r requirements_nl.in
six==1.17.0
    # via python-dateutil
smartypants==2.0.2
    # via notifications-utils
statsd==4.0.1
    # via notifications-utils
texttable==1.7.0
    # via pyexcel
urllib3==2.5.0
    # via
    #   botocore
    #   requests
    #   sentry-sdk
werkzeug==3.1.3
    # via
    #   flask
    #   flask-login
wtforms==3.2.1
    # via flask-wtf
xlrd==2.0.2
    # via pyexcel-xls
xlwt==1.3.0
    # via pyexcel-xls
